git-svn-id: branches/laksen/spc32@33690 -

+++ packages/gtk2/examples/scribble_simple/scribble_simple.pas
@@ -24,7 +24,7 @@ begin
                             -1);
   gdk_draw_rectangle (pixmap,
                               widget^.style^.white_gc,
-                              gTRUE,
+                              gint(gTRUE),
                               0, 0,
                               widget^.allocation.width,
                               widget^.allocation.height);
@@ -61,7 +61,7 @@ begin
 
   gdk_draw_rectangle (pixmap,
                               widget^.style^.black_gc,
-                          gTRUE,
+                          gint(gTRUE),
                               update_rect.x, update_rect.y,
                               update_rect.width, update_rect.height);
 
@@ -86,7 +86,7 @@ var
   state : TGdkModifierType;
 
 begin
-  if event^.is_hint = gTRUE then
+  if event^.is_hint = gint(gTRUE) then
     gdk_window_get_pointer (event^.window, @x, @y, @state)
   else begin
     x := round (event^.x);
